+++ app/javascript/engine/actions/FireAction.ts
@@ -505,6 +505,7 @@ export default class FireAction extends BaseAction {
         if (hitRoll.result > hitCheck) {
           if (needDice) { hitRoll.description += "hit" }
           anims.push({ loc: to, type: "hit" })
+          let infantryHit = false
           targets.forEach(t => {
             if (t.x === c.x && t.y === c.y) {
               if (t.counter.unit.isVehicle && !t.counter.unit.armored) {
@@ -546,7 +547,10 @@ export default class FireAction extends BaseAction {
                 this.game.moraleChecksNeeded.push({
                   unit: t.counter.unit, from: fcoords, to: c, incendiary: firing0.unit.incendiary
                 })
-                anims.push({ loc: to, type: "hit" })
+                if (!infantryHit) {
+                  anims.push({ loc: to, type: "hit" })
+                  infantryHit = true
+                }
               }
             }
           })