+++ app/javascript/routes/ProtectedRoute.tsx
@@ -35,7 +35,7 @@ export function ProtectedRoute() {
             localStorage.removeItem("validationNeeded")
             if (body.proto) { localStorage.setItem("proto", "true") }
             if (body.mcp) { localStorage.setItem("mcp", "true") }
-            if (path.includes("debug") && !body.proto) {
+            if (path.includes("/debug") && !body.proto) {
               navigate("/", { replace: true })
             }
           }